requirements.txt
opencv-python~=4.5.1.48
aiogram~=2.11.2
python-dotenv~=0.15.0
numpy~=1.20.0
Shapely~=1.7.1